create a useAddUserForm hooks to cleaner project
